Restore to end of learnopengl getting started
