+++ CMakeLists.txt
@@ -1,4 +1,5 @@
-cmake_minimum_required(VERSION 3.0)
+cmake_minimum_required(VERSION 3.11)
+set(CMAKE_CXX_STANDARD 11)
 project(glitch_game)
 
 include(${CMAKE_BINARY_DIR}/conanbuildinfo.cmake)
