+++ include/glitch/shader.h
@@ -1,10 +1,8 @@
-
-
 #ifndef SHADER_H
 #define SHADER_H
 
 #include <glad/glad.h>
-#include <glm/ext.hpp>
+#include <glm/glm.hpp>
 
 #include <string>
 #include <fstream>
@@ -77,12 +75,6 @@ public:
     { 
         glUseProgram(ID); 
     }
-    // deactivate the shader
-    // ------------------------------------------------------------------------
-    void deactivate()
-    {
-        glDeleteProgram(ID);
-    }
     // utility uniform functions
     // ------------------------------------------------------------------------
     void setBool(const std::string &name, bool value) const
@@ -99,13 +91,10 @@ public:
     { 
         glUniform1f(glGetUniformLocation(ID, name.c_str()), value); 
     }
-    void setVec4(const std::string &name, glm::vec4 value)
-    {
-		glUniform4fv(glGetUniformLocation(ID, name.c_str()), 1, glm::value_ptr(value));
-    }
-    void setMat4(const std::string &name, glm::mat4 value)
-    {
-		glUniformMatrix4fv(glGetUniformLocation(ID, name.c_str()), 1, GL_FALSE, glm::value_ptr(value));
+    // ------------------------------------------------------------------------
+    void setMat4(const std::string &name, glm::mat4 value) const
+    { 
+        glUniformMatrix4fv(glGetUniformLocation(ID, name.c_str()), 1, GL_FALSE, glm::value_ptr(value)); 
     }
 
 private:
@@ -135,5 +124,4 @@ private:
         }
     }
 };
-
 #endif
