--- src/lib/game/game.hpp
@@ -1,34 +1,0 @@
-#ifndef GAME_H
-#define GAME_H
-
-#include <vector>
-
-#include "world.hpp"
-#include "player.hpp"
-
-enum GameState {
-    Menu,
-    Play
-};
-
-// Class that handles all game logic
-class Game {
-  public:
-    Game(World first_world):
-      game_state_(Play),
-      player_(Player(first_world.startPosition(), first_world.startDirection())),
-      world_(first_world) {}
-
-    void run() {}
-
-    std::vector<Block> blocks() {
-      return world_.blocks();
-    }
-
-  private:
-    GameState game_state_;
-    Player player_;
-    World world_;
-};
-
-#endif
