--- src/shaders/f_color.glsl
@@ -1,9 +1,0 @@
-#version 330 core
-out vec4 FragColor;  
-
-uniform vec4 color;
-
-void main()
-{
-    FragColor = color;
-}
