+++ src/shaders/fragment.glsl
@@ -1,12 +1,11 @@
 #version 330 core
-out vec4 FragColor;  
-
-in vec3 ourColor;
+out vec4 FragColor;
+  
 in vec2 TexCoord;
 
 uniform sampler2D texture1;
 uniform sampler2D texture2;
-  
+
 void main()
 {
     FragColor = mix(texture(texture1, TexCoord), texture(texture2, TexCoord), 0.2);
